\name{ssmrob-package}
\alias{ssmrob-package}
%\alias{ssmrob}
\docType{package}
\title{
Robust Estimation and Inference in Sample Selection Models
}
\description{
Package provides a set of tools for robust estimation and inference for
models with sample selectivity.
}
\details{
\tabular{ll}{
Package: \tab robHeck \cr
Type: \tab Package \cr
Version: \tab 0.2 \cr
Date: \tab 2013-09-19 \cr
License: \tab GPL-2 \cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Mikhail Zhelonkin, Marc G. Genton, Elvezio Ronchetti

Maintainer: Mikhail Zhelonkin <Mikhail.Zhelonkin@unige.ch>
}
\references{
Heckman, J.J. (1979). Sample Selection Bias as a Specification Error. \emph{Econometrica}, 47, p. 153-161.

Zhelonkin, M., Genton M.G., and Ronchetti, E. (2013). Robust Inference in Sample Selection Models, \emph{Manuscript}.
}
\keyword{ Robust estimation }
\keyword{ Sample selection }
\seealso{
\code{\link[sampleSelection]{selection}}
}
%\examples{
%~~ simple examples of the most important functions ~~
%}

